bot-api/requirements.txt
discord.py
python-dotenv
selenium
requests
webdriver_manager